# src/main/resources/archetype-resources/__rootArtifactId__-exposition/src/main/resources/api.yml
swagger: "2.0"
info:
  description: "To complete"
  version: 1.0.0
  title: "${rootAggregate} overview services"

  contact:
    name: "France canaux CRC - Plateau Delivery  "
    email: "LIST.PFT.FRANCECANAUXCRCTEAMDEV@bnpparibas-pf.com"
host: "localhost"
basePath: "/"
...
schemes:
  - "https"
paths:
  /v1/${rootAggregateLowerCase}s:
    get:
      tags:
        - "${rootAggregateLowerCase}-v1"
      summary: "Get all the ${rootAggregate}s"
      operationId: "getAll${rootAggregate}s"
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/${rootAggregate}'
        500:
          description: "Internal Server Error"
          schema:
            $ref : '#/definitions/Error'
      deprecated: false
...
definitions:
...
  Error:
    type: object
    title: "Error"
    description: "Error structure"
    properties:
      additional_information:
        type: object
        description: "Additional data associated to this error (typically for automated process). String key/value format."
        properties:
          additionalProperties:
            type: string
      error:
        type: string
        description: "The error code. Ex : invalid_request. To use in automated process."
      error_description:
        type: string
        description: "The detailed human readable ASCII text explaining in details the error."
       
  # Statut du service 
...